Price accepts dot and not comma. Already added comma in price in db is fixed using cast(replace(price)) in query

+++ lib/model/finance_chart_data.dart
@@ -7,7 +7,7 @@ class Finance_Chart_Item{
 
   String date = "";
   String type = "";
-  int? amount = 0;
+  double? amount = 0;
 
   Finance_Chart_Item(
       {
